Adding formatted error message to fatal logs, #946
